Fixed Bootstrap icon version

+++ ci/bootstrap.py
@@ -1,18 +1,8 @@
 import urllib.request
-import bs4
 import zipfile
 import os
 
-_BASE = 'https://github.com'
-_URL = 'https://github.com/twbs/icons/releases/latest/'
-
-
-def _get_link():
-    resp = urllib.request.urlopen(_URL)
-    content = resp.read().decode(encoding="utf-8")
-    docTree = bs4.BeautifulSoup(content)
-    link = docTree.find_all('a', rel="nofollow")[0]['href']
-    return _BASE + link
+_URL = 'https://github.com/twbs/icons/releases/download/v1.4.1/bootstrap-icons-1.4.1.zip'
 
 
 def _download(url):
@@ -25,18 +15,19 @@ def _download(url):
         elif item[-3:] == "svg":
             archive.extract(item, os.path.join('overrides', '.icons'))
 
-    os.rename(
-        os.path.join('overrides', '.icons', dirname),
-        os.path.join('overrides', '.icons', 'bootstrap')
-    )
+    if not os.path.exists(os.path.join('overrides', '.icons', 'bootstrap')):
+        os.rename(
+            os.path.join('overrides', '.icons', dirname),
+            os.path.join('overrides', '.icons', 'bootstrap')
+        )
 
     archive.close()
     os.remove('temp.zip')
 
 
 def get_icon():
-    _download(_get_link())
+    _download(_URL)
 
 
 if __name__ == "__main__":
-    _download(_get_link())
+    _download(_URL)
